sdm845-common: rootdir: Enable zram deduplication Reference: https://blog.csdn.net/qq_38135221/article/details/127042461#t14 https://github.com/PixelExperience-Devices/kernel_oneplus_sdm845/commit/1af14fd9bd6f364c6d0511e4c27cfba3296b8c7a https://dl.acm.org/doi/10.1145/2797022.2797023
Android is one of the biggest users to use zram as swap and it's
really important to save amount of memory usage. There is a paper
that reports that duplication ratio of Android's memory content is
rather high. And, there is a similar work on zswap that also
reports that experiments has shown that around 10-15% of pages
stored in zswp are duplicates and deduplicate them provides some
benefits.